mail.local.c: cast off_t to `long long' for snprintf

+++ contrib/libexec/mail.local/mail.local.c
@@ -249,7 +249,8 @@ retry:
 	}
 
 	curoff = lseek(mbfd, 0, SEEK_END);
-	(void)snprintf(biffmsg, sizeof biffmsg, "%s@%lld\n", name, curoff);
+	(void)snprintf(biffmsg, sizeof biffmsg, "%s@%lld\n", name,
+	    (long long)curoff);
 	if (lseek(fd, 0, SEEK_SET) == (off_t)-1) {
 		mwarn("temporary file: %s", strerror(errno));
 		goto bad;
